Biztosan törli a foglalást? funkció.

+++ virtual_receptionist/Views/FormBooking.cs
@@ -102,21 +102,27 @@ namespace virtual_receptionist.Views
         {
             if (dataGridViewBookings.SelectedRows.Count > 0)
             {
-                object id = dataGridViewBookings.SelectedRows[0].Cells[0].Value;
-                object guest = dataGridViewBookings.SelectedRows[0].Cells[1].Value;
-                object room = dataGridViewBookings.SelectedRows[0].Cells[2].Value;
-                object numberOfGuests = dataGridViewBookings.SelectedRows[0].Cells[3].Value;
-                object arrivalDate = dataGridViewBookings.SelectedRows[0].Cells[4].Value;
-                object departureDate = dataGridViewBookings.SelectedRows[0].Cells[5].Value;
-                object paid = dataGridViewBookings.SelectedRows[0].Cells[6].Value;
+                DialogResult delete = MessageBox.Show("Biztosan törli a kijelölt foglalást?", "",
+                    MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
-                int row = dataGridViewBookings.SelectedRows[0].Index;
-                dataGridViewBookings.Rows.RemoveAt(row);
+                if (delete == DialogResult.Yes)
+                {
+                    object id = dataGridViewBookings.SelectedRows[0].Cells[0].Value;
+                    object guest = dataGridViewBookings.SelectedRows[0].Cells[1].Value;
+                    object room = dataGridViewBookings.SelectedRows[0].Cells[2].Value;
+                    object numberOfGuests = dataGridViewBookings.SelectedRows[0].Cells[3].Value;
+                    object arrivalDate = dataGridViewBookings.SelectedRows[0].Cells[4].Value;
+                    object departureDate = dataGridViewBookings.SelectedRows[0].Cells[5].Value;
+                    object paid = dataGridViewBookings.SelectedRows[0].Cells[6].Value;
 
-                controller.DeleteBooking(id, guest, room, numberOfGuests, arrivalDate, departureDate, paid);
+                    int row = dataGridViewBookings.SelectedRows[0].Index;
+                    dataGridViewBookings.Rows.RemoveAt(row);
 
-                MessageBox.Show("A foglalás sikeresen törlésre került!", "", MessageBoxButtons.OK,
-                    MessageBoxIcon.Information);
+                    controller.DeleteBooking(id, guest, room, numberOfGuests, arrivalDate, departureDate, paid);
+
+                    MessageBox.Show("A foglalás sikeresen törlésre került!", "", MessageBoxButtons.OK,
+                        MessageBoxIcon.Information);
+                }
             }
             else
             {
